+++ actix-web-lab/src/extract.rs
@@ -3,10 +3,6 @@
 /// An alias for [`actix_web::web::Data<T>`] with a more descriptive name.
 pub type SharedData<T> = actix_web::web::Data<T>;
 
-#[allow(deprecated)]
-pub use crate::body_hash::BodyHash;
-#[allow(deprecated)]
-pub use crate::body_hmac::{BodyHmac, HmacConfig};
 pub use crate::body_limit::BodyLimit;
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
 pub use crate::lazy_data::LazyData;