+++ actix-web-lab/src/lib.rs
@@ -26,8 +26,6 @@
 
 mod acceptable;
 mod body_extractor_fold;
-mod body_hash;
-mod body_hmac;
 mod body_limit;
 mod cache_control;
 mod catch_panic;